+++ docs/js/metro.js
@@ -3760,10 +3760,11 @@ window.METRO_LOCALES = {
         options: {
             target: false,
             shuffle: false,
-            length: false
+            length: false,
+            keys: ['1', '2', '3', '4', '5', '6', '7', '8', '9', '0']
         },
 
-        _keys: ['1', '2', '3', '4', '5', '6', '7', '8', '9', '0'],
+        //_keys: ['1', '2', '3', '4', '5', '6', '7', '8', '9', '0'],
 
         _create: function () {
             var that = this, element = this.element, o = this.options;
@@ -3778,6 +3779,10 @@ window.METRO_LOCALES = {
                 }
             });
 
+            if (typeof o.keys === 'string') {
+                o.keys = o.keys.split(",");
+            }
+
             this._createKeypad();
 
             element.data('keypad', this);
@@ -3786,21 +3791,24 @@ window.METRO_LOCALES = {
 
         _shuffleKeys: function(keypad){
             var that = this, element = this.element, o = this.options;
-            var keys = this._keys.slice(0);
+            var keys = o.keys.slice(0);
             var keypad = this._keypad;
+            var keys_length = keys.length + 2;
 
             if (o.shuffle) {
                 keys = keys.shuffle();
             }
 
-            keypad.html('');
+            keypad.html('').css({
+                width: keys_length / 4 * 32 + (keys_length / 4 + 1) * 2 + 2
+            });
 
             keys.map(function(i){
                 $("<div/>").addClass('key').html(i).data('key', i).appendTo(keypad);
             });
 
-            $("<div/>").addClass('key').html('<').data('key', '<').appendTo(keypad);
-            $("<div/>").addClass('key').html('C').data('key', 'C').appendTo(keypad);
+            $("<div/>").addClass('key').html('&larr;').data('key', '&larr;').appendTo(keypad);
+            $("<div/>").addClass('key').html('&times;').data('key', '&times;').appendTo(keypad);
         },
 
         _createKeypad: function(){
@@ -3861,16 +3869,16 @@ window.METRO_LOCALES = {
 
                 if (o.target) {
 
-                    if (parseInt(key.data('key')) >= 0) {
+                    if (key.data('key') !== '&larr;' && key.data('key') !== '&times;') {
                         if (o.length && $(o.target).val().length === o.length) {
                             return false;
                         }
                         $(o.target).val($(o.target).val() + '' + key.data('key'));
                     } else {
-                        if (key.data('key') === 'C') {
+                        if (key.data('key') === '&times;') {
                             $(o.target).val('');
                         }
-                        if (key.data('key') === '<') {
+                        if (key.data('key') === '&larr;') {
                             var val = $(o.target).val();
                             $(o.target).val(val.substring(0, val.length - 1))
                         }