+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 3.0.4
 * fixed float for dataTables
+* add keypad widget
 
 ## 3.0.3
 * New widget [Presenter](http://metroui.org.ua/presenter.html)